requirements.txt
django==5.1.7
python-dotenv==1.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
